Move ReadCids into entry package - Make ReadCids available to clients of this library - If test data files are not found, then skip test
